requirements.txt
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.10.0
dataclasses-json==0.5.7
fonttools==4.39.3
idna==2.10
kiwisolver==1.4.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
onemetric==0.1.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
roboflow==1.0.5
seaborn==0.12.2
six==1.16.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wget==3.2
clip @ git+https://github.com/openai/CLIP.git